+++ CHANGELOG.rst
@@ -46,6 +46,9 @@ Changelog
   :class:`~cryptography.hazmat.primitives.ciphers.aead.AESSIV`, and
   :class:`~cryptography.hazmat.primitives.ciphers.aead.ChaCha20Poly1305` to
   allow encrypting directly into a pre-allocated buffer.
+* Added a ``decrypt_into`` method to
+  :class:`~cryptography.hazmat.primitives.ciphers.aead.AESSIV` to allow
+  decrypting directly into a pre-allocated buffer.
 
 .. _v46-0-3: